Implement Ollama service.
This service requires configuration in the application.properties file:
`ollama.api.url` is the URL of the Ollama service and `ollama.model` is the name of the model.

When the server starts it will ping the Ollama service to make sure it is live and will ask it to pull the configured model.